meta: reproduce the gossip problem
This is the setup. Connecting to ldk-node -> okx (lnd) and try
to look at my node inside the network graph that runs core lightning.

My node is not big but it has 5/6 channels so it should be a problem.
After two days of fetching from okx I was not able to find my node
inside the gossip map.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>